Backport setting feature bit if sortpath is set. This helps with running 'bk cp' with bk-5.x and pulling that into bk-4.x and having it fail. Riders: + CRANKS almost always hang on 'bits' because of the bk-4.0.2d bkd startup.   Change the timing around to see if that helps. + Add something to save folks time figuring out why cranks failed.   Blocking cranks attempted from a repo with SCCS directories.
bk: 4e28f49cBBgZBJizKdjkQQWh-IjrrQ

+++ src/t/t.cp
@@ -21,7 +21,9 @@ do
 	echo "This is file 1, rev 1.$i" > file1
 	bk delta $Q -y"comment $i" file1
 done
+grep -q sortkey BitKeeper/log/features && fail has sortkey
 bk cp file1 file2 2> /dev/null
+grep -q sortkey BitKeeper/log/features || fail no sortkey
 if bk _test ! -f SCCS/s.file2; then echo Failed to create an s.file; exit 1; fi
 bk get $Q file1 file2 
 diff file1 file2
